Move saveResponsesToDatabase() to a separate file
